set StatusBar to not show by default
I was very surprised to see this bar when Surfingkeys updated.

I thought there was a bug or something. It even stays on when Chrome is in fullscreen mode. I feel like it should be off by default and that people can turn it on in the settings if they want.