Hold shift while typing hint opens hint in new tab
If any part of a hint is typed in as uppercase (i.e. with the shift key
held down), open the hint in a new tab.

Fixes #392

My rationale for this change:

This functionality was described by @brookhong in issue #4, in reference
to "open[ing] the selected hint in a new tab":

> just hold shift when you press the hint chars, actually you don't need to hold shift from the beginning. Holding shift at the last char is good to go.

This change should behave as described above.

This is also the standard behavior for other
Chrome Vim extensions like cVim and Vimium, implying that it's a
generally well accepted behavior.