SearchEngine auto-completion suggestions: Prevent extra HTTP requests (#396)
* SearchEngine suggestions: Prevent extra HTTP reqs

Previously, an HTTP request would be dispatched for each keystroke in an
OmniBar SearchEngine with a suggestionURL. This can lead to rate-limiting,
extra network traffic, and is unfriendly to API servers.

I've updated the search suggestion logic to use a timeout of 200ms
before dispatching an HTTP request. If another keystroke occurs before
this 200ms timeout, the previous request is canceled and the new one
again receives a 200ms timeout.

Example - When typing "testing" into a SearchEngine OmniBar with a
suggestionURL with less than 200ms between keystrokes:

Old: 7 HTTP requests are dispatched
New: 1 HTTP request is dispatched

Note: It may be desirable to make the timeout duration configurable.
Currently it's hard-coded as 200ms.

* Fix spacing

* Make Omnibar suggestion timeout configurable

Omnibar suggestion timeout is configurable via
settings.omnibarSuggestionTimeout and defaults to 200.

Updated README.md to reflect this. README_CN.md has not been updated (I
don't speak the language).

* Add comma

* Simplify language