Add runtime.conf.hintGroupStart
This option controls which hint group is active first.

Valid values are:
  - `first`: The first hint group is active first
	- `middle`: The middle hint group is active first
	- `last`: The last hint group is active first

If you notice you're often selecting hints located near the middle of
the page, it may be worthwhile to set `hintGroupStart` to `middle`.

Setting it to `middle` also has the effect of decreasing the maximum
potential number of number of cycles to reach any hint from `n` to
`n/2` where `n` is the number of groups.