Fix unnecessary scrolling when cursor is too small in Firefox
Since Firefox doesn't implement `scrollIntoViewIfNeeded`,
`isElementPartiallyInViewport` is used which considers elements smaller
than 4x4 pixels not visible. Thus when the cursor in visual mode
contains text smaller than that (e.g. the letter "i"), `scrollIntoView`
will always be called even if the cursor is currently visible.

This adds an optional `ignoreSize` argument to `scrollIntoViewIfNeeded`
to ignore the element size and is only used for the cursor.

Fixes another issue in #621