Disable autoFocus blurring when in insert or passthrough mode (#1137)
enableAutoFocus prevents the page from grabbing focus by blurring any
focus events. However, when we are in insert/passthrough mode, we
ignore intentional blurs by the user and (on some sites) blur the
active element while typing.

This patch changes this behavior so this auto-blurring will not take
place in insert or passthrough mode, as those modes are meant for
user input and/or unimpeded interaction with the page.