Fix enableAutoFocus blurring when entering insert mode dynamically (#1218)
Insert mode can be entered at the end of this function, if the cursor
somehow is focused via previous actions, but we are in normal mode. In
this case, we enter insert mode with passFocus=true, which makes the
first focus upon leaving insert mode pass through, even if
enableAutoFocus is set to false.

This fixes that issue by running passFocus = true only when we
actually focus an element.

I am assuming Mode.handleMapKey.call(self, event) does NOT need
passFocus = true, as it is used above without passFocus being set.