add link to wiki and FAQ in README.md (#1899)
It's not obvious from the github interface that there's an active wiki (other than the wiki tab appearing on the repo page), so it would be helpful to specifically mention both the FAQ and wiki in the README.md. I never would have found it if it weren't for someone pointing it out to me in a comment on an issue.