fix(uiframe): delay shadow DOM creation close #2253
The Amazon Prime Gaming page was failing to load
when SurfingKeys was enabled due to premature shadow DOM initialization.
his patch delays the creation of the shadow root until the DOM is ready,
ensuring that the page loads correctly.
Note that while this fixes the loading problem,
some commands (e.g., scrolling) still do not work.